+++ api/services/services.go
@@ -6,10 +6,6 @@ import (
 
 // Module exports services present
 var Module = fx.Options(
-	fx.Provide(NewStorageBucketService),
 	fx.Provide(NewUserService),
 	fx.Provide(NewTwilioService),
-	fx.Provide(NewGmailService),
-	fx.Provide(NewS3BucketService),
-	fx.Provide(NewGCPBillingService),
 )
